Move the test back to controllers package, implement the configs for testing
