+++ kafka-rest/src/test/java/io/confluent/kafkarest/controllers/ControllersModuleTest.java
@@ -1,30 +1,16 @@
-/*
- * Copyright 2021 Confluent Inc.
- *
- * Licensed under the Confluent Community License (the "License"); you may not use
- * this file except in compliance with the License.  You may obtain a copy of the
- * License at
- *
- * http://www.confluent.io/confluent-community-license
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations under the License.
- */
-
-package io.confluent.kafkarest.config;
+package io.confluent.kafkarest.controllers;
 
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertSame;
 
 import io.confluent.kafka.schemaregistry.client.SchemaRegistryClient;
-import io.confluent.kafkarest.controllers.ControllersModule;
-import io.confluent.kafkarest.controllers.SchemaRecordSerializer;
-import io.confluent.kafkarest.controllers.SchemaRecordSerializerThrowing;
+import io.confluent.kafkarest.config.ConfigModule.AvroSerializerConfigs;
+import io.confluent.kafkarest.config.ConfigModule.JsonschemaSerializerConfigs;
+import io.confluent.kafkarest.config.ConfigModule.ProtobufSerializerConfigs;
 import java.util.HashMap;
 import java.util.Map;
 import org.easymock.EasyMock;
+import org.glassfish.hk2.api.AnnotationLiteral;
 import org.glassfish.hk2.api.ServiceLocator;
 import org.glassfish.hk2.api.TypeLiteral;
 import org.glassfish.hk2.utilities.ServiceLocatorUtilities;
@@ -32,7 +18,16 @@ import org.glassfish.hk2.utilities.binding.AbstractBinder;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-public class SingletonSerializerTest {
+public class ControllersModuleTest {
+  private static final class AvroSerializerConfigsImpl
+      extends AnnotationLiteral<AvroSerializerConfigs> implements AvroSerializerConfigs {}
+
+  protected static final class JsonschemaSerializerConfigsImpl
+      extends AnnotationLiteral<JsonschemaSerializerConfigs>
+      implements JsonschemaSerializerConfigs {}
+
+  protected static final class ProtobufSerializerConfigsImpl
+      extends AnnotationLiteral<ProtobufSerializerConfigs> implements ProtobufSerializerConfigs {}
 
   private ServiceLocator serviceLocator;
 
@@ -51,13 +46,13 @@ public class SingletonSerializerTest {
             dummyConfig.put("schema.registry.url", "http://localhost:8081");
 
             bind(dummyConfig)
-                .qualifiedBy(new ConfigModule.AvroSerializerConfigsImpl())
+                .qualifiedBy(new AvroSerializerConfigsImpl())
                 .to(new TypeLiteral<Map<String, Object>>() {});
             bind(dummyConfig)
-                .qualifiedBy(new ConfigModule.JsonschemaSerializerConfigsImpl())
+                .qualifiedBy(new JsonschemaSerializerConfigsImpl())
                 .to(new TypeLiteral<Map<String, Object>>() {});
             bind(dummyConfig)
-                .qualifiedBy(new ConfigModule.ProtobufSerializerConfigsImpl())
+                .qualifiedBy(new ProtobufSerializerConfigsImpl())
                 .to(new TypeLiteral<Map<String, Object>>() {});
 
             install(new ControllersModule());
